*: Refactor new plan work flow
