+++ optimizer/preprocess.go
@@ -1,0 +1,29 @@
+// Copyright 2015 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package optimizer
+
+import (
+	"github.com/juju/errors"
+	"github.com/pingcap/tidb/ast"
+	"github.com/pingcap/tidb/context"
+	"github.com/pingcap/tidb/infoschema"
+)
+
+// Preprocess does preprocess work for optimizer.
+func Preprocess(node ast.Node, info infoschema.InfoSchema, ctx context.Context) error {
+	if err := ResolveName(node, info, ctx); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
